+++ app/_shared/activitypub/apSchemaService/index.ts
@@ -1,0 +1,1 @@
+export * as apSchemaService from "./service";
