+++ app/users/[userKey]/inbox/_services/inboxService/handlers/delete.ts
@@ -1,11 +1,11 @@
-import { inboxDeleteSchema } from "@/_shared/schema/delete";
+import { apSchemaService } from "@/_shared/activitypub/apSchemaService";
 import { prisma } from "@/_shared/utils/prisma";
 
 import { ActivitySchemaValidationError } from "./errors";
 import type { InboxHandler } from "./shared";
 
 export const handle: InboxHandler = async (activity) => {
-  const parsedDelete = inboxDeleteSchema.safeParse(activity);
+  const parsedDelete = apSchemaService.deleteSchema.safeParse(activity);
   if (!parsedDelete.success) {
     return new ActivitySchemaValidationError(parsedDelete.error, activity);
   }
